general: prepare for logic based on events
The generation of the full rows at the same time (1 period, the file period) is not suitable for adding AI. We need some kind of session state for each event and device so we can move the state gradually from the expected normal distribution and so on. So it is better to generate line by line each event. Much slower, but better results later.