python: ise folders as "modules"
Since py module system is a little weird compared to other languages, we can still keep folder separation of concerns instead of adding everithng to the same folder. Tried with python3.4+namespaces but it seems that we need to wait of real namespace resolution in next python version. So for now use the __init_ file and in the future we'll se if we change the file imports