time: MAJOR, configurable time speed to adapt sensor timestamps
Now, the time is a non-real-time variable, the SENSOR datetime will be every second, every minute every 10 mins etc, while the LOG FILE is still created every 10 mins. To allow users to get the file before 10 mins, the file is generated as soon as the app is running, then wait 10mins for another file generation.